requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
kafka-python
pandas
pyarrow
python-dotenv
